[Messages] Address a CHECK in MessageWrapper on activity destruction
When WebContents is destroyed as part of the activity destruction,
MessageWrapper hits a CHECK if a message is displayed. The reason for
the check is that the message wasn't dismissed.

This CL ensures that the SavePassword message is dismissed on the
delegate destruction. Additionally this CL prevents native
DismissMessage call from propagating to java when WebContents is being
destroyed.

BUG=1174656
R=lazzzis@chromium.org,twellington@chromium.org

Change-Id: I87101659f0bbbf604f92ae3456e83bed1fdba1ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2677321
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Pavel Yatsuk <pavely@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850961}