Remove RenderViewCreated use from ExtensionHost.
Prefer to watch for RenderFrame changes instead of RenderView changes.

This removes all reference to RenderViewHost from ExtensionHost and
updates classes that use ExtensionHost accordingly. All references are
actually used to get to the main frame, so we update everything to
pass RenderFrameHost of the main frame instead. Method names and
variable names are changed to drop the RenderView term as well.

This also removes the RenderViewReady override, but at time of
RenderFrameCreated it's too early to notify extensions about the
HOST being created as they expect RenderProcessHost::IsReady() to be
true for the renderer, and that comes later than when we are creating
the renderer-side objects. So we wait for RenderProcessHost to become
ready explicitly in ExtensionHost.

R=alexmos@chromium.org, reillyg@chromium.org

Bug: 763548
Change-Id: I34d1b34570ee2603533c0cd6e5a7a92178ed556b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2673502
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Sergey Volk <servolk@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850963}