Only allow a VM to write filenames to exo which are already shared
By setting a file in drag-and-drop or onto the clipboard, it will
automatcially be shared with the VM of any app that receives it.
We must only allow VMs to set paths which are already shared with them.

Bug: b/162602338
Change-Id: I0c9efb8fcc472e04083bedecec5db17373d71b54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2658537
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Jason Lin <lxj@google.com>
Auto-Submit: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850965}