Ensure parent-child consistency and improve CSS visibility updates
Adding a DCHECK when parent child relationships are inconsistent
shows a problem when CSS visibility changes. The DCHECK triggers
because the child tree for the hidden object was not recomputed,
and a "removed from tree" object was still considered an
unignored child of its parent. To fix this, LayoutObject now
calls ChildrenChanged() for the object with the visibility change,
which also caused beneficial de-duping of children-add events

Bug: None
Change-Id: I0d48d68d14c0e61a3cf89e5419433214641eb2ee
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2674082
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850967}