Implement scaled frame support in the Browser process.
This CL updates
VideoCaptureDeviceClient::OnIncomingCapturedExternalBuffer() to take
scaled video frames as a second parameter and pass them along to the
VideoFrameReceiver.

VideoFrameReceiver is implemented by the VideoCaptureController in the
Browser process. See https://photos.app.goo.gl/MWpR1RsmmZXsGSX27 for
overview of the pipeline that frames pass through. The controller is
updated to support multiple frames.

After this CL, scaled frames are correctly handled within the Browser
process and in previous CLs we added support in the Renderer process.
After this CL, all we need to update is the VideoFrameHandler mojom
interface and we'll have scaled frame support all the way from the
Capturer to the Renderer. Then all that remains is updating the
platform-specific code to produce and pass along scaled video frames.

video_capture_controller_unittest.cc is updated to test scaled frame
support, however because OnIncomingCapturedExternalBuffer()'s use of
external gfx::GpuMemoryBufferHandles only having been implemented on
macOS, this test can only be exercised on that platform even though the
VideoCaptureController code does not care about platform in and of
itself.

Bug: chromium:1157072
Change-Id: I6abec7c940ed17d0a129754ec3a1f78038bba80f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2649212
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851001}