Privacy Sandbox: Only enable if at least one API is available
To enable the independent launch of the Privacy Sandbox Settings,
without risking showing a user a UI to control features which are not
enabled, an additional layering is applied to inspecting feature state.

This additional layer ensures that at least one of the privacy sandbox
controlled APIs is available.

Bug: 1152336
Change-Id: I3ae85d61d8a9f22c073fcb4151d66daf350a08fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2675903
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Sean Harrison <harrisonsean@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/master@{#851047}