Disable AudioSandboxEnabledTest for lacros chrome.
crrev.com/c/2655075 which intended to disable audio sandbox for lacros
did not actually disable it since lacros is still built with OS_LINUX
set.

This patch ensures that that audio sandbox is disabled before and after
the macro flip. In addition to it, it removes tests related audio
sandbox.

Bug: 1171629,1168481
Change-Id: I0214592ba14d4f840011826b5c2ec84d7425428a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2674607
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Yuta Hijikata <ythjkt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851051}