Avoid circular dependency between system_info and system_storage
The only reason system_info depends on sytem_storage seems to be that
we need to access systeminfo::BuildStorageUnitInfo. This CL moves that
function to the system_info target to avoid a dependency on
system_storage.

This prepares for enabling 'gn check' for
extensions/browser/api/system_*.

Bug: 1158976, 1158972
Change-Id: I767e3bfacc499e1cf28aec2d487bd97cd0af53f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2671942
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Leon Han <leon.han@intel.com>
Cr-Commit-Position: refs/heads/master@{#851067}