Implement scaled video frame support by the VideoFrameHandler.
With this CL, scaled frames can be passed from the Capturer Process to
the Browser process (and thanks to prior CLs they will continue to be
passed all the way to the Renderer process).

This CL updates video_frame_handler.mojom and updates implementations.
BroadcastingReceiver is updated to support multiple frames and a basic
unit test is added to exercise the new path.

// Skipping because of purple bot, all other bots are green.
NOTRY=True

Bug: chromium:1157072
Change-Id: I8938589ecbe08a8fa0b009302a6b46dd30e44ffb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2671765
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851077}