Update Dice Web Signin Intercept to check for Guest windows.
Dice Web Signin Intercept is updated to check for Guest windows instead
of Guest profiles to decide if a Guest option should be offered or not.

This feature uses Ephemeral Guest profiles. If a Guest window is opened
and closed on an Ephemeral Guest profile, the profile is marked for
deletion and will not be used if another Guest window is opened. So in
the edge case that a Guest profile exist and a Guest window does not
exist, we are safe to use that Guest profile since it is not used to
open a window before.

Bug: 1157764
Change-Id: I1e17844a22d9cdacd5ef73a23925d6de65ffd6d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2676187
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851083}