Fix DCHECK crash in PasswordGenerationManager::PresaveGeneratedPassword.
This CL fixes a crash in a case when the password form that contains
a generated password is cleared dynamically and then the user taps on
a field that had generated password in the past, but is already empty by
that time.

[SharedPasswordController checkIfSuggestionsAvailableForForm:] =>
PasswordManager::UpdateStateOnUserInput(generation field, “”) =>
PasswordFormManager::UpdateStateOnUserInput(generation field, “”) =>
PasswordFormManager::PresaveGeneratedPasswordInternal(“”) =>
PasswordSaveManagerImpl::PresaveGeneratedPassword(“”) =>
PasswordGenerationManager::PresaveGeneratedPassword(“”) =>
DCHECK on empty string

The fix is to stop updating state on user input if there is no user
input.

Bug: 1174424
Change-Id: Icd264c47e7c3966f3b51eb2d306ebb4b42eaae97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2675938
Reviewed-by: Vadym Doroshenko  <dvadym@chromium.org>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/master@{#851084}