Bluetooth internals: Refactor tests
Refactor bluetooth_internals tests, to:
(1) Be in a separate folder, instead of at the top folder of
    chrome/test/data/webui
(2) Put mocha suites in their own file
(3) Put browser proxy classes and test utils in their own file.

This is in preparation for migrating bluetooth internals to JS modules,
as JS module tests cannot be in the same file that is passed as a source
to js2gtest.

Bug: 1173575
Change-Id: I0b49f1136bb5a2e95b2f819cb51655adf5f985f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2677501
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851149}