Improve QUIC UDP read error handling
Before this CL, our QUIC code would close the connection
if it encountered a zero-length UDP packet (which is valid)
or a UDP packet that was bigger than our receive buffer.
Since both can happen because of middleboxes, we should
just ignore them. This CL does just that.

This CL also adds a bool return type to
QuicChromiumPacketReader::Visitor::OnReadError. It's not
used by this CL but will be used in a subsequent CL to
allow the QUIC session to ignore some specific errors.

Note that some tests were incorrectly using a MockRead
of 0 (or OK which is defined to 0) as a way to signal
end-of-file and close the connection. While that works
with TCP, it is incorrect with UDP, and has been changed
to use ERR_CONNECTION_CLOSED instead.

Change-Id: I6ab6ab8d9dd103acfc0d76782d4ec0e0be89db4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2670035
Auto-Submit: David Schinazi <dschinazi@chromium.org>
Reviewed-by: Renjie Tang <renjietang@chromium.org>
Commit-Queue: David Schinazi <dschinazi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851217}