Native IO: Tighten access to NativeIOFileHost::SetLength().
NativeIOFileHost::SetLength() was introduced to work around a sandboxing
limitation on macOS < 10.15, which results in ftruncate() (the syscall
underlying base::File::SetLength()) always failing in renderer
processes. The workarond consisted of moving the ftruncate() call to the
browser process, and having the renderer invoke it via IPC.

This CL tightens access to SetLength() as follows.

1. The renderer calls ftruncate() directly on macOS >= 10.15 and on all
   other platforms. The NativeIOFileHost::SetLength() mojo IPC is only
   used on macOS < 10.15. The code path used is decided dynamically on
   macOS.  On non-Mac platforms, only the ftruncate() code is compiled
   into the binary.

2. NativeIOFileHost::SetLength() is only declared on macOS. Attempting
   to use it on other platforms introduces compilation errors.

3. On the browser side, NativeIOFileHost::SetLength() is only honored on
   on macOS < 10.15. On macOS >= 10.15, the implementation immediately
   calls mojo::ReportBadMessage().

The tightening described above limits the possibility for corrupted
renderers to use NativeIOFileHost::SetLength() to call ftruncate() on
arbitrary file descriptors to macOS < 10.15.

Bug: 914488
Change-Id: I38001cca64b51801bbc64de54c1c3fad8f16dbb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2660772
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851236}