Make "open all bookmarks" safe
chrome::OpenAll may prompt the user before opening many bookmarks. It
uses a nested RunLoop to do so. However, it takes as arguments
pointers that may be invalid at the end of this RunLoop.

This CL replaces chrome::OpenAll with chrome::OpenAllIfAllowed, which
returns immediately and opens the tabs asynchronously if prompting the
user. Instead of taking a content::PageNavigator pointer directly, it
takes a callback to fetch the pointer after the user acknowledges the
prompt. This can ensure a valid PageNavigator is used.

Another function chrome::OpenAllImmediately is added for tests which
shouldn't open a message box.

Fixed: 1161144
Change-Id: I6d0d73ec1d9deaf3cb339dc9646d7fe77a27674e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2650689
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Auto-Submit: Collin Baker <collinbaker@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851279}