Prevent tabstrip layout at wrong size on tab add
See attached bug.

Before this patch, it was possible that during tab add but before the
add animation, a preferred size could be calculated for the tabstrip.
Because the tabstrip itself was not subsequently invalidated after the
animation was started, the preferred size is briefly too large by one
tab width, resulting in the NTB being painted in the wrong spot for a
frame or two.

This change invalidates the layout after starting the animation, so
that the layout has to take into account the new preferred size (which
in turn takes into account the animating tab) on the next layout pass.

Bug: 1172910
Change-Id: Ibc472e712f06fb23917406916151358ced150420
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2674321
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Charlene Yan <cyan@chromium.org>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851289}