Reland Convert media/blink/webmediaplayer_impl from base::CancelableClosure to base::CancelableOnceClosure
My first attempt at this change resulted in some flaky tests on Linux.

The webmediaplayer_impl unit tests were previously using the
IsCancelled() api on CancelableCallback to determine is the disabling of
the video track was pending. This led to a slight mismatch between the
name of the function and what the function was actually doing since
it caused `IsDisableVideoTrackPending()` to return true even if the task
was not pending.

Now, since the `DisableVideoTrackIfNeeded()` method was posted to the
main task runner with a delay and PostDelayedTask actually ignores the
delay in unit tests my change resulted in flaky tests where the tests
would pass or fail depending on how fast the task runner would get to
execute the "delayed" task.

This was not a problem before as `IsDisableVideoTrackPending()` would
return true whether the task was pending or had already run. To keep
that behavior the test now check that the video track is either already
disabled or if it is pending being disabled.

Bug: 1170511
Change-Id: I23cf365f040e5303c751eb5a1bfc4a3bceb91f27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2676528
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Alex St-Onge <alstonge@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851297}