Enable the EduCoexistenceV2 flag by default
Enabling this feature by default required skipping it when other
OOBE browsertests are running.  Because the EDU Coexistence screen
is shown  immediately after signin for child users,  it conflicted
with the expected post-signin state that many OOBE tests expect
for child users.  Many of these tests wait for the disappearance
of the OOBE user creation screen.

Instead of special casing all these tests to expect a different
screen in the event of a child user (which would be irrelevant
to what they are trying to test), in this CL we always skip
the EDU Coexistence screen in OOBE tests.

The EDU Coexistence V2 screen itself is
provided by a web app loaded over the internet (similar to the
signin pages, which we also skip in these tests).  This makes
it difficult to test without some kind of web app mock.

In the future, we may consider adding a local mocked EDU
Coexistence flow server to support a test of the EDU
Coexistence screen itself separately from the other screens.


Bug: 1147681
Change-Id: I7a551658d8db2931913b789a329b1cdd70504e3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2673785
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Dan S <danan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851308}