fido: update the privacy notice for requests with rk="preferred"
(Also plumb the ResidentKeyRequirement through the
TransportAvailabilityInfo callback rather than its own
SetMightCreateResidentKey() interface because it's less boilerplatey.)

When creating a resident credential, we inform the user about the
privacy implication by displaying the string "A record of your visit to
this site will be kept on your security key." in WebAuthn UI. This was
accurate when "requireResidentKey" was a boolean value; but with enum
valued "residentKey" and its "preferred" case, whether a resident key is
created depends on the type of security key the user ultimately taps.
Hence, this replaces "will" with "may" for the "preferred" case.

We would be able to make this more accurate in cases where a second tap
is required after successful PIN entry, because at that point we do know
whether the chosen key supports discoverable credentials and (possibly)
whether its storage is exhausted. But updating the string at that stage
of the request would be extremely subtle and the user benefit marginal,
and it would never work with fingerprint reader keys, so I don't think
it's worth the additional complexity.

Fixed: 1152529
Change-Id: I5a36e46535fd1b5eb695ecbbe605d31420952f1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2677265
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Auto-Submit: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851326}