Fix DCHECK failure in biquad
When computing the tail time, we weren't correctly handling the case
where the poles are +/- 1.  This causes a DCHECK failure in clampTo
because the values are NaN.

Now, check for poles on or outside the unit circle and return
infinity for the tail frame.

One test is added for this.  A debug build without this CL fails.
With this, there is no DCHECK failure.  Also ran the test case from
issue 1170531 with a debug build and there are no biquad DCHECK
failures.

Bug: 1172332
Test: BiquadFilter/tail-frame-peaking.html
Change-Id: I0cc449b4b703794006a62bea7343b108e6127236
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2675092
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851332}