[WebLayer] Add Browser.setDarkModeStrategy API.
This CL adds a Browser.setDarkModeStrategy API that allows embedders to
control how WebLayer renders pages when in dark mode. Note that this
doesn't provide a way for embedders to decouple a page's dark mode with
WebLayer's entirely (e.g. render a page in dark mode if WebLayer itself
is in light, or vice versa). This can be added as a separate setting if
we need it in the future.

Bug: 1136653
Change-Id: I0003f1e095fda3280a840cf794619050b48fd581
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2674202
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851359}