Update system_display_interface
The system_display_interface was out of sync with the extensions
API causing a closure compile error when migrating the device page
to Polymer 3.

This CL adds the missing method to the interface and adds a readme
for how to generate or update these interfaces.

In the case pf system_display_iterface.js the generated file is
manually modified to replace System.display with SystemDisplay
to avoid a separate closure compile error.

Bug: 1045266
Change-Id: Iedda334aa968b27f125bc0b28a448ca5665695e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2668368
Commit-Queue: Claude van der Merwe <cvandermerwe@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851391}